Support more TOTP credentials.
The `uint8_t` could store a maximum offset of 255, which would represent
around 13 credentials with a key size of 20 bytes. I ran into this limitation.
The `uint16_t` allows for more than 3000 credentials and some other limit of the
watch's memory is probably reached before that point.

+++ movement/watch_faces/complication/totp_face.h
@@ -60,7 +60,7 @@ typedef struct {
     uint8_t steps;
     uint32_t current_code;
     uint8_t current_index;
-    uint8_t current_key_offset;
+    uint16_t current_key_offset;
 } totp_state_t;
 
 void totp_face_setup(movement_settings_t *settings, uint8_t watch_face_index, void ** context_ptr);
